# docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Blog API",
        "description": "Open API v3.0 documentation for the Blog API",
        "contact": {
            "email": "git@ahowell.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Blog API Server"
        }
    ],
    "paths": {
        "/comments/{id}": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "Gets a single comment by comment id",
                "description": "Returns a single comment by comment id, including the id of the post it relates to and the user that made it",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Comment id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CommentResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Resource not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Comments"
                ],
                "summary": "Updates an existing comment by comment id",
                "description": "Updates an existing comment by comment id and then returns the comment, including the id of the\n    post it relates to and the user that made it",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Comment id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CommentUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CommentResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "422": {
                        "description": "Unprocessable entity"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Comments"
                ],
                "summary": "Deletes a comment by comment id",
                "description": "Deletes an existing comment by comment id and returns no content",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Comment id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
        "/posts/{id}": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Gets a single post by post id",
                "description": "Returns a single post by post id along with user that created it and any associated tags",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Post id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Resource not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Posts"
                ],
                "summary": "Updates an existing post by post id",
                "description": "Updates an existing post by post id and any links to associated tags and returns the result. If\n    a tag id is not included in the update request, its association with the post will be removed",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Post id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PostRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "422": {
                        "description": "Unprocessable entity"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Posts"
                ],
                "summary": "Deletes a post by post id",
                "description": "Deletes a post by post id and returns no content",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Post id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
        "/tags": {
            "get": {
                "tags": [
                    "Tags"
                ],
                "summary": "Gets all tags",
                "description": "Returns all tags",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TagResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            },
            "post": {
                "tags": [
                    "Tags"
                ],
                "summary": "Creates a new tag",
                "description": "Creates a new tag and then returns the tag",
                "operationId": "store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TagRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TagResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Unprocessable entity"
                    }
                }
            }
        },
        "/tags/{id}": {
            "get": {
                "tags": [
                    "Tags"
                ],
                "summary": "Gets a single tag by tag id",
                "description": "Returns a single tag by tag id",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Tag id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TagResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Resource not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Tags"
                ],
                "summary": "Updates an existing tag by tag id",
                "description": "Updates an existing tag by tag id and returns the result",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Tag id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TagRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TagResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "422": {
                        "description": "Unprocessable entity"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tags"
                ],
                "summary": "Deletes a tag by tag id",
                "description": "Deletes a tag by tag id and returns no content",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Tag id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/tags/{id}/posts": {
            "get": {
                "tags": [
                    "Tags"
                ],
                "summary": "Gets a single tag by tag id and any posts related to it",
                "description": "Gets a single tag by tag id and any posts related to it",
                "operationId": "showWithPosts",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Tag id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TagWithPostsResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Resource not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CommentUpdateRequest": {
                "properties": {
                    "body": {
                        "schema": "CommentUpdateRequest",
                        "description": "Body of the post",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "PostRequest": {
                "properties": {
                    "title": {
                        "schema": "PostRequest",
                        "description": "Post title",
                        "type": "string"
                    },
                    "body": {
                        "schema": "PostRequest",
                        "description": "Body of the post",
                        "type": "string"
                    },
                    "tags": {
                        "schema": "PostRequest",
                        "description": "Array of tag ids related to this post",
                        "type": "array",
                        "items": {
                            "type": "integer"
                        }
                    }
                },
                "type": "object"
            },
            "TagRequest": {
                "properties": {
                    "title": {
                        "schema": "TagRequest",
                        "description": "Tag title, max 32 characters",
                        "type": "string"
                    },
                    "color": {
                        "schema": "TagRequest",
                        "description": "Color of the tag in full rgb format including the leading # e.g: #FF00FF",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "CommentResource": {
                "properties": {
                    "id": {
                        "schema": "CommentResource",
                        "description": "Comment id",
                        "type": "integer",
                        "format": "int64"
                    },
                    "body": {
                        "schema": "CommentResource",
                        "description": "Body of the comment",
                        "type": "string"
                    },
                    "post_id": {
                        "schema": "CommentResource",
                        "description": "Id of the post this comment relates to",
                        "type": "integer",
                        "format": "int64"
                    },
                    "created_at": {
                        "schema": "CommentResource",
                        "description": "Date and time the comment was created",
                        "type": "string"
                    },
                    "updated_at": {
                        "schema": "CommentResource",
                        "description": "Date and time the comment was last updated",
                        "type": "string"
                    },
                    "author": {
                        "$ref": "#/components/schemas/UserResource"
                    }
                },
                "type": "object"
            },
...
            "PostResource": {
                "properties": {
                    "id": {
                        "schema": "PostResource",
                        "description": "Post id",
                        "type": "integer",
                        "format": "int64"
                    },
                    "title": {
                        "schema": "PostResource",
                        "description": "Post title",
                        "type": "string"
                    },
                    "body": {
                        "schema": "PostResource",
                        "description": "Body of the post",
                        "type": "string"
                    },
                    "created_at": {
                        "schema": "PostResource",
                        "description": "Date and time the post was created",
                        "type": "string"
                    },
                    "updated_at": {
                        "schema": "PostResource",
                        "description": "Date and time the post was last updated",
                        "type": "string"
                    },
                    "author": {
                        "$ref": "#/components/schemas/UserResource"
                    },
                    "tags": {
                        "schema": "PostResource",
                        "description": "Tags related to this post",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/TagResource"
                        }
                    },
                    "comment_count": {
                        "schema": "PostResource",
                        "description": "Number of comments related to this post",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "TagResource": {
                "properties": {
                    "id": {
                        "schema": "TagResource",
                        "description": "Tag id",
                        "type": "integer",
                        "format": "int64"
                    },
                    "title": {
                        "schema": "TagResource",
                        "description": "Tag title",
                        "type": "string"
                    },
                    "color": {
                        "schema": "TagResource",
                        "description": "Tag color",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "TagWithPostsResource": {
                "properties": {
                    "id": {
                        "schema": "TagWithPostsResource",
                        "description": "Tag id",
                        "type": "integer",
                        "format": "int64"
                    },
                    "title": {
                        "schema": "TagWithPostsResource",
                        "description": "Tag title",
                        "type": "string"
                    },
                    "color": {
                        "schema": "TagWithPostsResource",
                        "description": "Tag color",
                        "type": "string"
                    },
                    "posts": {
                        "schema": "TagWithPostsResource",
                        "description": "Posts associated with the tag",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PostResource"
                        }
                    }
                },
                "type": "object"
            },
            "UserResource": {
                "properties": {
                    "id": {
                        "schema": "UserResource",
                        "description": "User id",
                        "type": "integer",
                        "format": "int64"
                    },
                    "displayname": {
                        "schema": "UserResource",
                        "description": "Display name of the user",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    }
}